requirement.txt
fastapi
uvicorn
pydantic
google-generativeai
fastapi-cors